+++ types/ember-metrics/metrics-adapters/base.d.ts
@@ -1,0 +1,3 @@
+declare module 'ember-metrics/metrics-adapters/base' {
+    export default class BaseAdapter {}
+}